implementing workaround for root not-found with multiple root layouts
